# gen/go/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/proto/coordinator/v1/coordinator.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CoordinatorService"
    },
    {
      "name": "WhoisBotService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/bots/connect": {
      "post": {
        "summary": "Connect bot",
        "description": "Binds bot to establish connection to coordination server",
        "operationId": "CoordinatorService_BotConnect",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1BotConnectionInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Unique bot name.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "apiKey",
            "description": "Coordinator API key.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "CoordinatorService"
        ]
      }
    },
    "/v1/coordinator/probe": {
      "put": {
        "summary": "Start new probe",
        "description": "Creates new probing task with variable type",
        "operationId": "CoordinatorService_NewProbe",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GenericProbeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GenericProbe"
            }
          }
        ],
        "tags": [
          "CoordinatorService"
        ]
      }
    },
    "/v1/coordinator/probes": {
      "put": {
        "summary": "Start multiple new probes",
        "description": "Creates new probing tasks with variable types with stream",
        "operationId": "CoordinatorService_NewProbes",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/v1GenericProbeResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of v1GenericProbeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1GenericProbes"
            }
          }
        ],
        "tags": [
          "CoordinatorService"
        ]
      }
    },
    "/v1/status": {
      "get": {
        "summary": "Get coordinator status",
        "description": "Returns coordinator status including bot connections, queue stats and other metrics",
        "operationId": "CoordinatorService_GetStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CoordinatorState"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "CoordinatorService"
        ]
      }
    },
    "/v1/whois/probe": {
      "post": {
        "summary": "Probe host",
        "description": "ProbeHost immediately starts whois query (manual use). Production use is limited due to system passing new tasks through redis queues",
        "operationId": "WhoisBotService_ProbeHost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1WhoisInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "host.uuid",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "host.dn",
            "description": "Fully Qualified Domain Name (FQDN) or simply Domain Name (DN)",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "host.url",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "host.email",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "host.ip.v4",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "host.ip.v6",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "byte"
          }
        ],
        "tags": [
          "WhoisBotService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com. As of May 2023, there are no widely used type server\nimplementations and no plans to implement one.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1BotConnectionInfo": {
      "type": "object",
      "properties": {
        "queuePool": {
          "type": "string",
          "description": "Redis queue pool name."
        },
        "queuePassword": {
          "type": "string",
          "description": "Redis queue pool password."
        }
      },
      "description": "Connection credentials for bot to connect to the Redis queue and start listening for new tasks from coordinator."
    },
    "v1CoordinatorInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Coordination server unique name."
        },
        "mode": {
          "$ref": "#/definitions/v1CoordinatorMode"
        }
      }
    },
    "v1CoordinatorMode": {
      "type": "string",
      "enum": [
        "COORDINATOR_MODE_MASTER",
        "COORDINATOR_MODE_WORKER"
      ],
      "default": "COORDINATOR_MODE_MASTER",
      "description": "Master mode enables child coordinators to be connected. Worker mode is always connected to master coordinator and accepts new tasks",
      "title": "Work mode"
    },
    "v1CoordinatorState": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/v1CoordinatorInfo"
        }
      }
    },
    "v1GenericProbe": {
      "type": "object",
      "properties": {
        "UUID": {
          "type": "string",
          "description": "Unique probe UUID"
        },
        "type": {
          "$ref": "#/definitions/v1ProbeType",
          "description": "Previously defined probe type"
        },
        "responseRequired": {
          "type": "boolean",
          "description": "Is sync response required"
        },
        "whois": {
          "$ref": "#/definitions/v1ProbeOptions"
        }
      }
    },
    "v1GenericProbeResponse": {
      "type": "object",
      "properties": {
        "UUID": {
          "type": "string"
        },
        "whois": {
          "$ref": "#/definitions/v1WhoisInfo"
        }
      }
    },
    "v1GenericProbes": {
      "type": "object",
      "properties": {
        "probes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GenericProbe"
          }
        }
      }
    },
    "v1Host": {
      "type": "object",
      "properties": {
        "uuid": {
          "type": "string"
        },
        "dn": {
          "type": "string",
          "title": "Fully Qualified Domain Name (FQDN) or simply Domain Name (DN)"
        },
        "url": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "ip": {
          "$ref": "#/definitions/v1IPAddress"
        }
      }
    },
    "v1IPAddress": {
      "type": "object",
      "properties": {
        "v4": {
          "type": "integer",
          "format": "int64"
        },
        "v6": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "IPv4 or IPv6 IP address\nhttps://stackoverflow.com/questions/51395679/what-is-the-most-efficient-protobuf-type-in-c-for-storing-ipv4-or-ipv6-addre"
    },
    "v1ProbeOptions": {
      "type": "object",
      "properties": {
        "host": {
          "$ref": "#/definitions/v1Host",
          "description": "Host to be scanned"
        }
      }
    },
    "v1ProbeType": {
      "type": "string",
      "enum": [
        "PROBE_TYPE_UNKNOWN",
        "PROBE_TYPE_WHOIS",
        "PROBE_TYPE_PING"
      ],
      "default": "PROBE_TYPE_UNKNOWN"
    },
    "v1WhoisInfo": {
      "type": "object",
      "properties": {
        "host": {
          "type": "string"
        },
        "rawData": {
          "type": "string"
        },
        "sourceServer": {
          "type": "string"
        },
        "redirected": {
          "$ref": "#/definitions/v1WhoisInfo",
          "title": "redirected is used to link to data sourced from original whois server"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}